backend/requirements.txt
Flask
flask-cors
pandas
scikit-learn
matplotlib
seaborn
numpy
biopython
wordcloud
pillow
lxml
requests
plotly
kaleido
